+++ SharpSploit/Execution/Win32.cs
@@ -338,6 +338,21 @@ namespace SharpSploit.Execution
                 SidTypeComputer = 9
             }
 
+            [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
+            public struct SHARE_INFO_1
+            {
+                public string shi1_netname;
+                public uint shi1_type;
+                public string shi1_remark;
+
+                public SHARE_INFO_1(string netname, uint type, string remark)
+                {
+                    this.shi1_netname = netname;
+                    this.shi1_type = type;
+                    this.shi1_remark = remark;
+                }
+            }
+
             [DllImport("netapi32.dll")]
             public static extern int NetLocalGroupEnum(
                 [MarshalAs(UnmanagedType.LPWStr)] string servername,
@@ -382,6 +397,17 @@ namespace SharpSploit.Execution
                 int prefmaxlen,
                 out int entriesread,
                 out int totalentries,
+                ref int resume_handle
+            );
+
+            [DllImport("Netapi32.dll", CharSet = CharSet.Unicode)]
+            public static extern int NetShareEnum(
+                StringBuilder ServerName,
+                int level,
+                ref IntPtr bufPtr,
+                uint prefmaxlen,
+                ref int entriesread,
+                ref int totalentries,
                 ref int resume_handle
             );
 
@@ -663,6 +689,47 @@ namespace SharpSploit.Execution
             [DllImport("advapi32.dll", SetLastError = true)]
             public static extern Boolean RevertToSelf();
 
+            [DllImport("advapi32.dll", SetLastError = true)]
+            public static extern IntPtr OpenSCManager(
+                string machineName,
+                string databaseName,
+                SCM_ACCESS dwAccess
+            );
+
+            [DllImport("advapi32.dll", SetLastError = true)]
+            public static extern IntPtr OpenService(
+                IntPtr hSCManager,
+                string lpServiceName,
+                SERVICE_ACCESS dwDesiredAccess
+            );
+
+            [DllImport("advapi32.dll", SetLastError = true)]
+            public static extern IntPtr CreateService(
+                IntPtr hSCManager,
+                string lpServiceName,
+                string lpDisplayName,
+                SERVICE_ACCESS dwDesiredAccess,
+                SERVICE_TYPE dwServiceType,
+                SERVICE_START dwStartType,
+                SERVICE_ERROR dwErrorControl,
+                string lpBinaryPathName,
+                string lpLoadOrderGroup,
+                string lpdwTagId,
+                string lpDependencies,
+                string lpServiceStartName,
+                string lpPassword
+            );
+
+            [DllImport("advapi32.dll", SetLastError = true)]
+            public static extern bool DeleteService(
+                IntPtr hService
+            );
+
+            [DllImport("advapi32.dll", SetLastError = true)]
+            public static extern bool CloseServiceHandle(
+                IntPtr hSCObject
+            );
+
             //https://msdn.microsoft.com/en-us/library/windows/desktop/ms682434(v=vs.85).aspx
             [Flags]
             public enum CREATION_FLAGS
@@ -701,6 +768,152 @@ namespace SharpSploit.Execution
                 LOGON32_PROVIDER_WINNT35,
                 LOGON32_PROVIDER_WINNT40,
                 LOGON32_PROVIDER_WINNT50
+            }
+
+            [Flags]
+            public enum SCM_ACCESS : uint
+            {
+                SC_MANAGER_CONNECT = 0x00001,
+                SC_MANAGER_CREATE_SERVICE = 0x00002,
+                SC_MANAGER_ENUMERATE_SERVICE = 0x00004,
+                SC_MANAGER_LOCK = 0x00008,
+                SC_MANAGER_QUERY_LOCK_STATUS = 0x00010,
+                SC_MANAGER_MODIFY_BOOT_CONFIG = 0x00020,
+
+                SC_MANAGER_ALL_ACCESS = ACCESS_MASK.STANDARD_RIGHTS_REQUIRED |
+                    SC_MANAGER_CONNECT |
+                    SC_MANAGER_CREATE_SERVICE |
+                    SC_MANAGER_ENUMERATE_SERVICE |
+                    SC_MANAGER_LOCK |
+                    SC_MANAGER_QUERY_LOCK_STATUS |
+                    SC_MANAGER_MODIFY_BOOT_CONFIG,
+
+                GENERIC_READ = ACCESS_MASK.STANDARD_RIGHTS_READ |
+                    SC_MANAGER_ENUMERATE_SERVICE |
+                    SC_MANAGER_QUERY_LOCK_STATUS,
+
+                GENERIC_WRITE = ACCESS_MASK.STANDARD_RIGHTS_WRITE |
+                    SC_MANAGER_CREATE_SERVICE |
+                    SC_MANAGER_MODIFY_BOOT_CONFIG,
+
+                GENERIC_EXECUTE = ACCESS_MASK.STANDARD_RIGHTS_EXECUTE |
+                    SC_MANAGER_CONNECT | SC_MANAGER_LOCK,
+
+                GENERIC_ALL = SC_MANAGER_ALL_ACCESS,
+            }
+
+            [Flags]
+            public enum ACCESS_MASK : uint
+            {
+                DELETE = 0x00010000,
+                READ_CONTROL = 0x00020000,
+                WRITE_DAC = 0x00040000,
+                WRITE_OWNER = 0x00080000,
+                SYNCHRONIZE = 0x00100000,
+                STANDARD_RIGHTS_REQUIRED = 0x000F0000,
+                STANDARD_RIGHTS_READ = 0x00020000,
+                STANDARD_RIGHTS_WRITE = 0x00020000,
+                STANDARD_RIGHTS_EXECUTE = 0x00020000,
+                STANDARD_RIGHTS_ALL = 0x001F0000,
+                SPECIFIC_RIGHTS_ALL = 0x0000FFFF,
+                ACCESS_SYSTEM_SECURITY = 0x01000000,
+                MAXIMUM_ALLOWED = 0x02000000,
+                GENERIC_READ = 0x80000000,
+                GENERIC_WRITE = 0x40000000,
+                GENERIC_EXECUTE = 0x20000000,
+                GENERIC_ALL = 0x10000000,
+                DESKTOP_READOBJECTS = 0x00000001,
+                DESKTOP_CREATEWINDOW = 0x00000002,
+                DESKTOP_CREATEMENU = 0x00000004,
+                DESKTOP_HOOKCONTROL = 0x00000008,
+                DESKTOP_JOURNALRECORD = 0x00000010,
+                DESKTOP_JOURNALPLAYBACK = 0x00000020,
+                DESKTOP_ENUMERATE = 0x00000040,
+                DESKTOP_WRITEOBJECTS = 0x00000080,
+                DESKTOP_SWITCHDESKTOP = 0x00000100,
+                WINSTA_ENUMDESKTOPS = 0x00000001,
+                WINSTA_READATTRIBUTES = 0x00000002,
+                WINSTA_ACCESSCLIPBOARD = 0x00000004,
+                WINSTA_CREATEDESKTOP = 0x00000008,
+                WINSTA_WRITEATTRIBUTES = 0x00000010,
+                WINSTA_ACCESSGLOBALATOMS = 0x00000020,
+                WINSTA_EXITWINDOWS = 0x00000040,
+                WINSTA_ENUMERATE = 0x00000100,
+                WINSTA_READSCREEN = 0x00000200,
+                WINSTA_ALL_ACCESS = 0x0000037F
+            }
+
+            [Flags]
+            public enum SERVICE_ACCESS : uint
+            {
+                SERVICE_QUERY_CONFIG = 0x00001,
+                SERVICE_CHANGE_CONFIG = 0x00002,
+                SERVICE_QUERY_STATUS = 0x00004,
+                SERVICE_ENUMERATE_DEPENDENTS = 0x00008,
+                SERVICE_START = 0x00010,
+                SERVICE_STOP = 0x00020,
+                SERVICE_PAUSE_CONTINUE = 0x00040,
+                SERVICE_INTERROGATE = 0x00080,
+                SERVICE_USER_DEFINED_CONTROL = 0x00100,
+
+                SERVICE_ALL_ACCESS = (ACCESS_MASK.STANDARD_RIGHTS_REQUIRED |
+                    SERVICE_QUERY_CONFIG |
+                    SERVICE_CHANGE_CONFIG |
+                    SERVICE_QUERY_STATUS |
+                    SERVICE_ENUMERATE_DEPENDENTS |
+                    SERVICE_START |
+                    SERVICE_STOP |
+                    SERVICE_PAUSE_CONTINUE |
+                    SERVICE_INTERROGATE |
+                    SERVICE_USER_DEFINED_CONTROL),
+
+                GENERIC_READ = ACCESS_MASK.STANDARD_RIGHTS_READ |
+                    SERVICE_QUERY_CONFIG |
+                    SERVICE_QUERY_STATUS |
+                    SERVICE_INTERROGATE |
+                    SERVICE_ENUMERATE_DEPENDENTS,
+
+                GENERIC_WRITE = ACCESS_MASK.STANDARD_RIGHTS_WRITE |
+                    SERVICE_CHANGE_CONFIG,
+
+                GENERIC_EXECUTE = ACCESS_MASK.STANDARD_RIGHTS_EXECUTE |
+                    SERVICE_START |
+                    SERVICE_STOP |
+                    SERVICE_PAUSE_CONTINUE |
+                    SERVICE_USER_DEFINED_CONTROL,
+
+                ACCESS_SYSTEM_SECURITY = ACCESS_MASK.ACCESS_SYSTEM_SECURITY,
+                DELETE = ACCESS_MASK.DELETE,
+                READ_CONTROL = ACCESS_MASK.READ_CONTROL,
+                WRITE_DAC = ACCESS_MASK.WRITE_DAC,
+                WRITE_OWNER = ACCESS_MASK.WRITE_OWNER,
+            }
+
+            [Flags]
+            public enum SERVICE_TYPE : uint
+            {
+                SERVICE_KERNEL_DRIVER = 0x00000001,
+                SERVICE_FILE_SYSTEM_DRIVER = 0x00000002,
+                SERVICE_WIN32_OWN_PROCESS = 0x00000010,
+                SERVICE_WIN32_SHARE_PROCESS = 0x00000020,
+                SERVICE_INTERACTIVE_PROCESS = 0x00000100,
+            }
+
+            public enum SERVICE_START : uint
+            {
+                SERVICE_BOOT_START = 0x00000000,
+                SERVICE_SYSTEM_START = 0x00000001,
+                SERVICE_AUTO_START = 0x00000002,
+                SERVICE_DEMAND_START = 0x00000003,
+                SERVICE_DISABLED = 0x00000004,
+            }
+
+            public enum SERVICE_ERROR
+            {
+                SERVICE_ERROR_IGNORE = 0x00000000,
+                SERVICE_ERROR_NORMAL = 0x00000001,
+                SERVICE_ERROR_SEVERE = 0x00000002,
+                SERVICE_ERROR_CRITICAL = 0x00000003,
             }
         }
 
@@ -1298,6 +1511,14 @@ namespace SharpSploit.Execution
                 int cb,
                 out int pSize
             );
+
+            [StructLayout(LayoutKind.Sequential)]
+            public struct UNICODE_STRING
+            {
+                public UInt16 Length;
+                public UInt16 MaximumLength;
+                public IntPtr Buffer;
+            }
 
             public struct PROCESS_BASIC_INFORMATION
             {