+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Is64BitProcess Host function (credit @TheWover)
 - Added GetProcessorArchitecture, GetParentProcess, GetProcessOwner, IsWow64, and supporting P/Invoke signatures (credit @rasta-mouse)
 - Added Keylogger class (credit @checkymander)
+- Added SCM class, PowerShellRemoting class, Host.GetDacl function (credit @rasta-mouse)
+- Added NetShareEnum functionality for Share Enumeration (credit @checkymander)
+- Added in-memory export parsing (credit @b33f)
 
 ### Changed
 - Improved DynamicInvoke library (credit @TheWover)